CYF 1.0 try n°1

+++ Assets/Tiled2Unity/Scripts/Editor/Tiled2UnityMenuItems.cs
@@ -5,6 +5,10 @@ using System.IO;
 using System.Linq;
 using System.Text;
 
+#if!UNITY_WEBPLAYER
+using System.Xml.Linq;
+#endif
+
 using UnityEditor;
 using UnityEngine;
 
@@ -12,17 +16,25 @@ namespace Tiled2Unity
 {
     class Tiled2UnityMenuItems
     {
-#if !UNITY_WEBPLAYER
+#if!UNITY_WEBPLAYER
         // Convenience function for packaging this library
         [MenuItem("Tiled2Unity/Export Tiled2Unity Library ...")]
         static void ExportLibrary()
         {
-            string name = String.Format("Tiled2Unity.{0}.unitypackage", ImportTiled2Unity.ThisVersion);
+            // Get the version from our Tiled2Unity.export.txt library data file
+            TextAsset textAsset = AssetDatabase.LoadAssetAtPath("Assets/Tiled2Unity/Tiled2Unity.export.txt", typeof(TextAsset)) as TextAsset;
+            XDocument xml = XDocument.Parse(textAsset.text);
+            string version = xml.Element("Tiled2UnityImporter").Element("Header").Attribute("version").Value;
+
+            // Export the package
+            string name = String.Format("Tiled2Unity.{0}.unitypackage", version);
             var path = EditorUtility.SaveFilePanel("Save Tiled2Unity library as unity package.", "", name, "unitypackage");
             if (path.Length != 0)
             {
                 List<string> packageFiles = new List<string>();
-                packageFiles.AddRange(EnumerateAssetFilesAt("Assets/Tiled2Unity", ".cs", ".shader", ".txt"));
+
+                // Export all C# files, shaders, text files, and some select materials
+                packageFiles.AddRange(EnumerateAssetFilesAt("Assets/Tiled2Unity",".cs", ".shader", ".txt", "t2uSprite-Depth.mat", "t2uSprite-DiffuseDepth.mat"));
                 AssetDatabase.ExportPackage(packageFiles.ToArray(), path);
             }
         }
@@ -39,11 +51,11 @@ namespace Tiled2Unity
         //    DeleteAssetsAt("Assets/Tiled2Unity/Textures");
         //}
 
-        private static IEnumerable<string> EnumerateAssetFilesAt(string dir, params string[] extensions)
+        private static IEnumerable<string> EnumerateAssetFilesAt(string dir, params string[] endPatterns)
         {
             foreach (string f in Directory.GetFiles(dir))
             {
-                if (extensions.Any(ext => String.Compare(ext, Path.GetExtension(f), true) == 0))
+                if (endPatterns.Any(pat => f.EndsWith(pat, true, null)))
                 {
                     yield return f;
                 }
@@ -51,7 +63,7 @@ namespace Tiled2Unity
 
             foreach (string d in Directory.GetDirectories(dir))
             {
-                foreach (string f in EnumerateAssetFilesAt(d, extensions))
+                foreach (string f in EnumerateAssetFilesAt(d, endPatterns))
                 {
                     yield return f;
                 }
@@ -71,6 +83,12 @@ namespace Tiled2Unity
 
                 // Just to be safe. Do not remove scripts.
                 if (f.EndsWith(".cs", true, null))
+                    continue;
+
+                // Do not remove special materials
+                if (f.EndsWith("t2uSprite-Depth.mat", true, null))
+                    continue;
+                if (f.EndsWith("t2uSprite-DiffuseDepth.mat", true, null))
                     continue;
 
                 AssetDatabase.DeleteAsset(f);
